feat: Rm unnecessary code, upd paths (#61)
