--- AIAgent/common/strings.py
@@ -1,4 +1,0 @@
-AV_COVERAGE_COL_NAME = "average coverage"
-MEDIAN_COVERAGE_COL_NAME = "median coverage"
-EUC_DIST2FULL_COV_COL_NAME = "euc dist to full coverage"
-COV_DEVIATION_COL_NAME = "coverage deviation"
