+++ AIAgent/connection/game_server_conn/messages.py
@@ -6,9 +6,8 @@ from typing import Optional
 
 from common.game import GameMap, GameState, Reward
 from config import FeatureConfig
+from connection.game_server_conn.unsafe_json import obj_from_dict
 from dataclasses_json import config, dataclass_json
-
-from .unsafe_json import obj_from_dict
 
 
 class ClientMessageType(str, Enum):
